error handling overhaul and more error checking
This combines a number of minor changes and one large
design change.

The basic usage pattern for requests looked like this:
	try:
	    result = client.query("...")
	    if result.ok:
	        # use result
	    else:
	        # handle error
	except:
	    # handle error

But as users, we don't inherently care about the nature of the error,
we care whether the operation succeeded or failed.

So this suggests two options; raise an exception for any error yielded
by the server, or trap exceptions and convert them into errors. I went
with the latter, because it makes the logic slightly easier for the
batch cases.

So, we drop the separate "exec" member of the result; error holds
the error, which may be an exception or may be a string we got back
from the server.

When returning a batch of results, we wrap them in a UserList type
which computes the set of errors (or None) and an overall "ok" value
since that's easier for users to check.

Test cases are variously updated for this.

We check more things at client creation, giving reasonable exceptions
for them. We require that hostport be set if you're providing an API
key, because otherwise you'll try to connect to a local featurebase
using SSL, which fails with a very strange error instead of failing
with a clear and comprehensible error.

We also handle a couple of other cases, such as an empty-string database
being set. (No database is fine, an empty string causes weird failures.)

We also address a couple of other issues. Our sample code in our
README had a bunch of stuff indented in a way that I think was an
accident. I've cleaned up that behavior a bit. We had one example
printing a string's ".description" member, which it doesn't have,
because it's just a string. Also, the "bulkInsert" example created
99 batches of 9,999 items, rather than 100 batches of 10,000 items,
because python ranges are half-open. Fixed that.

We also add a link to the documentation to our README (which I believe
is also used to generate our PyPI page).

Overall, this justifies a bump to 0.1.0. We've changed semantics,
and by strict semver perhaps we should bump the major version number,
but actually we are not yet ready to call this a 1.0.