refactor error handling, handle some config errors better
If you hit an error like a missing hostport, you end up getting
two screenfulls or so of stack traces, which tells you almost
nothing, on a "wrong SSL version" message, which tells you something
incorrect.

We handle some errors more directly. However, the deeper issue
is that it's super annoying to both have to check for errors AND
have to catch exceptions. We treat errors as exceptions now; if
the server yields a non-200 status code, or we get an error message
even on a 200, we raise the error as an exception.

The "stoponerrors" feature is now the only real option, because
it doesn't actually make *sense* to set it to "no" outside of
test cases.

This dramatically simplifies a lot of code.

In the async case, we still raise a GroupException of any exceptions
encountered, unless there's none in which case we yield results.

Some tests are a bit stricter or more careful; for instance, if we
do two queries, we can just assert that we get exactly two responses,
not "two or more".